Reflect changes with SteamKit. Still cannot log in yet
